Movement and Attack Bug Fix
In an actual build production, the movement would stutter constantly. In order to fix this, it needed to be placed in FixedUpdate(). That being said, the sword swing attack is then messed up as it would not register the attack button consistently. In order to fix that, the if statement with the attack was placed in Update() and then called in FixedUpdate(). Other areas were changed as well to get everything functioning properly.