Projectile Enemy, Shield, and General Bug Fixes
Another form of Octorok that can shoot rock projectiles that hit the player and give damage to them. Made a simple shield script that destroys the rock projectile if it hits the shield (couldn't figure out how to make it bounce of the shield). Fixed a bug where the heart and minimap canvas would sometimes show up even when paused.