Fixing Some Issues
Fixed a bug where when player held down the attack button while holding diagonally, player turned to Octorok. This was due to the fact that some of the animations placed in the attack state where of the Octorok. Also fixed an issue in regards to the previous commit where enemies wouldn't attack each other but player couldn't deal damage to them.