Projectile Octorok and Keaton Update
Apparently, the duplicate rock projectile bug wasn't completely fixed. With the updated script, in the hierarchy, the rock projectile still displays twice on the first initial launch, however, in-game, it functions how it should be. As well,  updated script for Keaton enemy to make move speed faster when player is in range, and updates to initial value when CheckDistance() is called again.